Remove GmailTestBase
